Combine and tidy two of the file_system_entry impls
This merges together ChefRepositoryFileSystemEntry and FileSystemEntry,
since they weren't really being used separately.
